+++ src/buffer/base/CoRE4INET_QueueBuffer.h
@@ -93,7 +93,7 @@ class QueueBuffer : public virtual Buffer
          * @return number of messages in the buffer
          *
          */
-        virtual unsigned int size() const;
+        virtual size_t size() const;
 
         /**
          * @brief Resets the buffer, deletes all messages
